@@ -24,13 +27,13 @@ android {
             storePassword localProperties['storePassword']
         }
     }
-    compileSdkVersion 30
+    compileSdkVersion 34
     defaultConfig {
         applicationId "pl.bydgoszcz.guideme.podlewacz"
         minSdkVersion 19
-        targetSdkVersion 30
-        versionCode 7
-        versionName "0.7"
+        targetSdkVersion 34
+        versionCode 8
+        versionName "0.8"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         vectorDrawables.useSupportLibrary true
         multiDexEnabled true
@@ -49,47 +52,46 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility = 1.8
-        targetCompatibility = 1.8
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
     }
 
     kotlinOptions {
-        jvmTarget = "1.8"
+        jvmTarget = "17"
     }
 }
 
 dependencies {
-    def lifecycle_version = "2.2.0"
-    def dagger_version = "2.20"
+    implementation 'com.google.firebase:firebase-crashlytics:18.4.0'
+    implementation 'com.google.firebase:firebase-analytics:21.3.0'
 
-    kapt "com.google.dagger:dagger-android-processor:$dagger_version"
-    kapt "com.google.dagger:dagger-compiler:$dagger_version"
+    kapt 'com.google.dagger:dagger-android-processor:2.47'
+    kapt 'com.google.dagger:dagger-compiler:2.47'
 
     implementation fileTree(include: ['*.jar'], dir: 'libs')
 
-    implementation "com.google.dagger:dagger-android:$dagger_version"
-    implementation "com.google.dagger:dagger-android-support:$dagger_version"
+    implementation 'com.google.dagger:dagger-android:2.47'
+    implementation 'com.google.dagger:dagger-android-support:2.47'
 
-    implementation "androidx.core:core:1.1.0"
-    implementation "androidx.recyclerview:recyclerview:1.1.0"
-    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
-    implementation "androidx.exifinterface:exifinterface:1.1.0"
-    implementation "androidx.work:work-runtime-ktx:2.3.0"
-    implementation "androidx.lifecycle:lifecycle-livedata:$lifecycle_version"
-    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_version"
-    implementation "androidx.coordinatorlayout:coordinatorlayout:1.1.0"
-    implementation "com.google.android.material:material:1.2.0-alpha04"
+    implementation "androidx.core:core-ktx:1.10.1"
+    implementation "androidx.recyclerview:recyclerview:1.3.1"
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
+    implementation "androidx.exifinterface:exifinterface:1.3.6"
+    implementation "androidx.work:work-runtime-ktx:2.8.1"
+    implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.6.1"
+    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.6.1"
+    implementation "androidx.coordinatorlayout:coordinatorlayout:1.2.0"
+    implementation "com.google.android.material:material:1.11.0-alpha02"
 
     implementation 'com.google.android.gms:play-services-gcm:17.0.0'
 
-    implementation 'com.crashlytics.sdk.android:crashlytics:2.10.1'
 
-    implementation 'com.google.firebase:firebase-core:17.2.2'
-    implementation 'com.google.firebase:firebase-perf:19.0.5'
+    implementation 'com.google.firebase:firebase-core:21.1.1'
+    implementation 'com.google.firebase:firebase-perf:20.4.0'
 
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.0'
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.0'
+    implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.9.0'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.7.3'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.3'
 
     implementation('com.squareup.picasso:picasso:2.71828') {
         exclude group: 'com.android.support', module: 'support-annotations'
