@@ -8,8 +8,8 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import kotlinx.android.synthetic.main.fragment_flower_item.view.*
 import pl.bydgoszcz.guideme.podlewacz.R
+import pl.bydgoszcz.guideme.podlewacz.databinding.FragmentFlowerItemBinding
 import pl.bydgoszcz.guideme.podlewacz.notifications.getElapsedTime
 import pl.bydgoszcz.guideme.podlewacz.notifications.getRemainingDaysMessage
 import pl.bydgoszcz.guideme.podlewacz.notifications.updateRemainingTime
@@ -24,21 +24,18 @@ class FlowerRecyclerViewAdapter(
         private val mContext: Context,
         private val mListener: OnListFragmentInteractionListener?)
     : RecyclerView.Adapter<FlowerRecyclerViewAdapter.ViewHolder>() {
-    private val mOnClickListener: View.OnClickListener
-    private var stopped = false
-
-    init {
-        mOnClickListener = View.OnClickListener { v ->
-            val position = v.tag as Int? ?: return@OnClickListener
-            val item = items[position]
-            mListener?.onListFragmentInteraction(item)
-        }
+    private val mOnClickListener: View.OnClickListener = View.OnClickListener { v ->
+        val position = v.tag as Int? ?: return@OnClickListener
+        val item = items[position]
+        mListener?.onListFragmentInteraction(item)
     }
+    private var stopped = false
+    private var _binding: FragmentFlowerItemBinding? = null
+    private val binding get() = _binding!!
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        val view = LayoutInflater.from(parent.context)
-                .inflate(R.layout.fragment_flower_item, parent, false)
-        return ViewHolder(view) { stopped }
+        _binding = FragmentFlowerItemBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+        return ViewHolder(binding) { stopped }
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
@@ -47,23 +44,23 @@ class FlowerRecyclerViewAdapter(
 
         item.updateRemainingTime()
 
-        holder.mNameView.text = Html.fromHtml(item.item.name)
+        binding.name.text = Html.fromHtml(item.item.name)
 
         if (showRemainingTime) {
-            holder.mFrequencyText.text = item.item.notification.getRemainingDaysMessage(mContext)
-            holder.mFrequencyText.visibility = View.VISIBLE
+            binding.remainingDaysTextView.text = item.item.notification.getRemainingDaysMessage(mContext)
+            binding.remainingDaysTextView.visibility = View.VISIBLE
         } else {
-            holder.mFrequencyText.visibility = View.GONE
+            binding.remainingDaysTextView.visibility = View.GONE
         }
 
         val color = if (item.remainingTime.toDays() < 0) android.R.color.holo_red_dark else R.color.colorPrimary
 
-        holder.mFrequencyText.setTextColor(mContext.resources.getColorFromResource(color))
-        holder.mView.tag = position
+        binding.remainingDaysTextView.setTextColor(mContext.resources.getColorFromResource(color))
+        binding.root.tag = position
 
         loadImage(holder, item, color, position)
 
-        holder.mView.setOnClickListener(mOnClickListener)
+        binding.root.setOnClickListener(mOnClickListener)
         holder.worker.onTick = {
             refresh(it)
         }
@@ -72,14 +69,14 @@ class FlowerRecyclerViewAdapter(
     data class ImageItemTag(val color: Int, val position: Int)
 
     private fun loadImage(holder: ViewHolder, item: UiItem, color: Int, position: Int) {
-        val imageItemTag = holder.mFlowerImageView.tag as ImageItemTag?
+        val imageItemTag = binding.flowerImage.tag as ImageItemTag?
         if (imageItemTag?.color == color && imageItemTag.position == position) {
             return
         }
         if (item.item.imageUrl.isEmpty()) {
             return
         }
-        holder.mFlowerImageView.tag = ImageItemTag(color, position)
+        binding.flowerImage.tag = ImageItemTag(color, position)
         ImageLoader.setImageWithCircle(holder.mFlowerImageView,
                 item.item.imageUrl,
                 onError = {
@@ -88,7 +85,7 @@ class FlowerRecyclerViewAdapter(
     }
 
     private fun refresh(holder: ViewHolder) {
-        val position = holder.mView.tag as Int? ?: return
+        val position = holder.binding.root.tag as Int? ?: return
         val item = items[position]
         if (!item.item.notification.enabled) {
             return
@@ -116,10 +113,10 @@ class FlowerRecyclerViewAdapter(
         stopped = false
     }
 
-    inner class ViewHolder(val mView: View, isStopped: () -> Boolean) : RecyclerView.ViewHolder(mView) {
-        val mNameView: TextView = mView.name
-        val mFlowerImageView: ImageView = mView.flowerImage
-        val mFrequencyText: TextView = mView.remainingDaysTextView
+    inner class ViewHolder(val binding: FragmentFlowerItemBinding, isStopped: () -> Boolean) : RecyclerView.ViewHolder(binding.root) {
+        val mNameView: TextView = binding.name
+        val mFlowerImageView: ImageView = binding.flowerImage
+        val mFrequencyText: TextView = binding.remainingDaysTextView
         val worker: Worker = Worker.constructAndRun(isStopped, this)
 
         override fun toString(): String {
